# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chemical Inventory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: getInventoryCounts
      tags: [inventory]
      summary: List inventory counts
      parameters:
        - name: storeId
          in: query
          schema:
            type: integer
        - name: weekOf
          in: query
          schema:
            type: string
            description: ISO date string for the week (YYYY-MM-DD, Monday)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of inventory counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryCount"
    post:
      operationId: submitInventoryCount
      tags: [inventory]
      summary: Submit a weekly inventory count
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitInventoryCountBody"
      responses:
        "201":
          description: Count submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryCount"
  /inventory/{countId}:
    get:
      operationId: getInventoryCount
      tags: [inventory]
      summary: Get a single inventory count
      parameters:
        - name: countId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryCount"
    patch:
      operationId: updateInventoryCount
      tags: [inventory]
      summary: Update notes or submittedBy on a count
      parameters:
        - name: countId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryCountBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
    delete:
      operationId: deleteInventoryCount
      tags: [inventory]
      summary: Delete an inventory count and its entries
      parameters:
        - name: countId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /inventory/scan:
    post:
      operationId: scanInventorySheet
      tags: [inventory]
      summary: Upload photo of paper count sheet and extract counts via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInventorySheetBody"
      responses:
        "200":
          description: Extracted inventory counts from the photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanInventorySheetResult"
...
  /push-tokens:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register or update a device push token (upsert by token value)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                platform:
                  type: string
                label:
                  type: string
                minSeverity:
                  type: string
              required: [token]
      responses:
        "200":
          description: Token updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushToken"
        "201":
          description: Token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushToken"
  /push-tokens/{token}:
    delete:
      operationId: unregisterPushToken
      tags: [notifications]
      summary: Remove a device push token registration
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /admin/push-tokens:
    get:
      operationId: getAdminPushTokens
      tags: [admin]
      summary: List all registered push notification devices (requires admin PIN header)
      parameters:
        - name: x-admin-pin
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of registered push tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PushToken"
  /admin/push-tokens/{tokenId}:
    delete:
      operationId: deleteAdminPushToken
      tags: [admin]
      summary: Delete a registered push token (requires admin PIN header)
      parameters:
        - name: x-admin-pin
          in: header
          required: true
          schema:
            type: string
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /admin/push-receipts:
    get:
      operationId: getAdminPushReceipts
      tags: [admin]
      summary: List push delivery receipts (requires admin PIN header)
      parameters:
        - name: x-admin-pin
          in: header
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, ok, error]
      responses:
        "200":
          description: List of push delivery receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PushReceipt"
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
      required:
        - success
        - id
...
    UpdateInventoryCountBody:
      type: object
      properties:
        submittedBy:
          type: string
        notes:
          type: string
          nullable: true
...
    InventoryEntry:
      type: object
      properties:
        chemicalId:
          type: integer
        chemicalName:
          type: string
        quantity:
          type: number
        unit:
          type: string
      required:
        - chemicalId
        - chemicalName
        - quantity
        - unit
    InventoryCount:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        storeName:
          type: string
        weekOf:
          type: string
          description: ISO date string for the Monday of the week
        submittedBy:
          type: string
        userId:
          type: integer
          nullable: true
        userName:
          type: string
          nullable: true
          description: Resolved employee name from the users table (null for legacy rows)
        notes:
          type: string
          nullable: true
        submittedAt:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/InventoryEntry"
      required:
        - id
        - storeId
        - storeName
        - weekOf
        - submittedBy
        - submittedAt
        - entries
    SubmitInventoryCountBody:
      type: object
      properties:
        storeId:
          type: integer
        weekOf:
          type: string
        submittedBy:
          type: string
        userId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        entries:
          type: array
          items:
            type: object
            properties:
              chemicalId:
                type: integer
              quantity:
                type: number
            required:
              - chemicalId
              - quantity
      required:
        - storeId
        - weekOf
        - submittedBy
        - entries
    ScanInventorySheetBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image of the paper count sheet
        chemicals:
          type: array
          description: List of chemicals to look for in the sheet
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              unit:
                type: string
            required:
              - id
              - name
              - unit
      required:
        - imageBase64
        - chemicals
    ScanInventorySheetResult:
      type: object
      properties:
        entries:
          type: array
          items:
            type: object
            properties:
              chemicalId:
                type: integer
              chemicalName:
                type: string
              quantity:
                type: number
              confidence:
                type: string
                enum: [high, medium, low]
            required:
              - chemicalId
              - chemicalName
              - quantity
              - confidence
        rawText:
          type: string
          description: Raw text extracted from the image
      required:
        - entries
        - rawText
...
    PushToken:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        platform:
          type: string
        label:
          type: string
          nullable: true
        minSeverity:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - token
        - platform
        - minSeverity
        - createdAt
        - updatedAt
    PushReceipt:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: string
        token:
          type: string
        alertId:
          type: integer
          nullable: true
        storeName:
          type: string
        chemicalName:
          type: string
        severity:
          type: string
        status:
          type: string
          enum: [pending, ok, error]
        errorCode:
          type: string
          nullable: true
        sentAt:
          type: string
        checkedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - token
        - storeName
        - chemicalName
        - severity
        - status
        - sentAt
        - createdAt